backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Shubham9455",email = "shubham.jaiswal.ece21@itbhu.ac.in"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
